refactor: Hilt 의존성 주입 구현
