+++ app/src/main/java/com/young/sportsmatch/data/source/remote/ApiClient.kt
@@ -1,14 +1,7 @@
 package com.young.sportsmatch.data.source.remote
 
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import com.young.sportsmatch.BuildConfig
 import com.young.sportsmatch.data.model.User
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.PUT
@@ -30,28 +23,4 @@ interface ApiClient {
         @Body user: User,
     ): Response<Map<String, String>>
 
-    companion object {
-
-        private const val BASE_URL = BuildConfig.FIRE_BASE_URL
-        private val moshi = Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .build()
-
-        fun create(): ApiClient {
-            val logger = HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BASIC
-            }
-
-            val client = OkHttpClient.Builder()
-                .addInterceptor(logger)
-                .build()
-
-            return Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .client(client)
-                .addConverterFactory(MoshiConverterFactory.create(moshi))
-                .build()
-                .create(ApiClient::class.java)
-        }
-    }
 }
